+++ js/src/jit/CodeGenerator.cpp
@@ -6995,7 +6995,7 @@ CodeGenerator::link(JSContext *cx, types::CompilerConstraintList *constraints)
     // This will reduce deviations due to the state of the machine at link-time, making it easier
     // to isolate the impact of changes to the input source code.
     char *buf = js_pod_malloc<char>(code->instructionsSize() * 4 + 1);
-    for (int i = 0; i < code->instructionsSize(); i++) {
+    for (size_t i = 0; i < code->instructionsSize(); i++) {
     	JS_snprintf(buf + 4 * i, 4, "\\x%02x", *(code->raw() + i));
     }
     buf[code->instructionsSize() * 4] = '\0';
